refactor: move file submenu to separate file
